#226 renamed and moved auth providers to separate package
