+++ src/main/java/lexek/wschat/security/social/provider/GoodGameAuthProvider.java
@@ -1,4 +1,4 @@
-package lexek.wschat.security.social;
+package lexek.wschat.security.social.provider;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.ImmutableList;
@@ -7,6 +7,9 @@ import com.google.common.net.UrlEscapers;
 import io.netty.util.CharsetUtil;
 import lexek.wschat.chat.e.InvalidStateException;
 import lexek.wschat.security.SecureTokenGenerator;
+import lexek.wschat.security.social.SocialProfile;
+import lexek.wschat.security.social.SocialRedirect;
+import lexek.wschat.security.social.SocialToken;
 import lexek.wschat.util.JsonResponseHandler;
 import lexek.wschat.util.RestResponse;
 import lexek.wschat.util.RestResponseHandler;
@@ -18,24 +21,20 @@ import org.apache.http.client.methods.HttpPost;
 import org.apache.http.message.BasicNameValuePair;
 
 import java.io.IOException;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 
-public class GoogleSocialAuthProvider implements SocialAuthProvider {
+public class GoodGameAuthProvider implements SocialAuthProvider {
     private final String name;
     private final String clientId;
     private final String clientSecret;
     private final String url;
     private final HttpClient httpClient;
     private final SecureTokenGenerator secureTokenGenerator;
-    private final String scopesString;
 
-    public GoogleSocialAuthProvider(
+    public GoodGameAuthProvider(
         String clientId,
         String clientSecret,
         String url,
-        Set<String> scopes,
         String name,
         HttpClient httpClient,
         SecureTokenGenerator secureTokenGenerator
@@ -46,15 +45,15 @@ public class GoogleSocialAuthProvider implements SocialAuthProvider {
         this.name = name;
         this.httpClient = httpClient;
         this.secureTokenGenerator = secureTokenGenerator;
-        this.scopesString = scopes.stream().collect(Collectors.joining(" "));
     }
 
     @Override
     public SocialRedirect getRedirect() {
         String state = secureTokenGenerator.generateRandomToken(32).replace("=", "");
-        String result = "https://accounts.google.com/o/oauth2/v2/auth?response_type=code&client_id=" + clientId +
+        String result = "http://api2.goodgame.ru/oauth/authorize?response_type=code" +
+            "&client_id=" + UrlEscapers.urlPathSegmentEscaper().escape(clientId) +
             "&redirect_uri=" + UrlEscapers.urlPathSegmentEscaper().escape(url) +
-            "&scope=" + UrlEscapers.urlPathSegmentEscaper().escape(scopesString) +
+            "&scope=chat.token" +
             "&state=" + UrlEscapers.urlPathSegmentEscaper().escape(state) +
             "&prompt=consent" +
             "&access_type=offline";
@@ -63,7 +62,7 @@ public class GoogleSocialAuthProvider implements SocialAuthProvider {
 
     @Override
     public SocialToken authenticate(String code) throws IOException {
-        HttpPost request = new HttpPost("https://www.googleapis.com/oauth2/v4/token");
+        HttpPost request = new HttpPost("http://api2.goodgame.ru/oauth");
         HttpEntity entity = new UrlEncodedFormEntity(ImmutableList.of(
             new BasicNameValuePair("client_secret", clientSecret),
             new BasicNameValuePair("client_id", clientId),
@@ -86,25 +85,25 @@ public class GoogleSocialAuthProvider implements SocialAuthProvider {
                 refreshToken
             );
         } else {
-            throw new InvalidStateException(rootNode.get("error_description").asText());
+            throw new InvalidStateException(rootNode.get("detail").asText());
         }
     }
 
     @Override
     public SocialProfile getProfile(SocialToken token) throws IOException {
-        HttpGet request = new HttpGet("https://www.googleapis.com/plus/v1/people/me");
+        HttpGet request = new HttpGet("http://api2.goodgame.ru/info");
         request.setHeader(HttpHeaders.AUTHORIZATION, "Bearer " + token.getToken());
         request.setHeader(HttpHeaders.ACCEPT, "application/json");
         JsonNode userData = httpClient.execute(request, JsonResponseHandler.INSTANCE);
-        String externalName = userData.get("displayName").asText().toLowerCase();
-        String externalId = userData.get("id").asText();
-        String email = userData.get("emails").get(0).get("value").asText();
-        return new SocialProfile(externalId, name, externalName, email, token);
+        JsonNode userNode = userData.get("user");
+        String externalName = userNode.get("username").asText().toLowerCase();
+        String externalId = userNode.get("user_id").asText();
+        return new SocialProfile(externalId, name, externalName, null, token);
     }
 
     @Override
     public SocialToken refresh(SocialToken oldToken) throws IOException {
-        HttpPost request = new HttpPost("https://www.googleapis.com/oauth2/v4/token");
+        HttpPost request = new HttpPost("http://api2.goodgame.ru/oauth");
         request.setHeader(HttpHeaders.ACCEPT, "application/json");
         request.setEntity(new UrlEncodedFormEntity(ImmutableList.of(
             new BasicNameValuePair("client_id", clientId),
@@ -114,12 +113,13 @@ public class GoogleSocialAuthProvider implements SocialAuthProvider {
         )));
         JsonNode response = httpClient.execute(request, JsonResponseHandler.INSTANCE);
         String token = response.get("access_token").asText();
+        String refreshToken = response.get("refresh_token").asText();
         long expiresIn = TimeUnit.SECONDS.toMillis(response.get("expires_in").asLong() - 60);
         return new SocialToken(
             name,
             token,
             System.currentTimeMillis() + expiresIn,
-            oldToken.getRefreshToken()
+            refreshToken
         );
     }
 
@@ -130,7 +130,7 @@ public class GoogleSocialAuthProvider implements SocialAuthProvider {
 
     @Override
     public boolean checkEmail() {
-        return true;
+        return false;
     }
 
     @Override
