+++ src/main/java/lexek/wschat/security/social/provider/VkAuthProvider.java
@@ -1,26 +1,23 @@
-package lexek.wschat.security.social;
+package lexek.wschat.security.social.provider;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.google.common.collect.ImmutableList;
 import com.google.common.net.HttpHeaders;
 import com.google.common.net.UrlEscapers;
-import io.netty.util.CharsetUtil;
 import lexek.wschat.chat.e.InvalidStateException;
 import lexek.wschat.security.SecureTokenGenerator;
+import lexek.wschat.security.social.SocialProfile;
+import lexek.wschat.security.social.SocialRedirect;
+import lexek.wschat.security.social.SocialToken;
 import lexek.wschat.util.JsonResponseHandler;
 import lexek.wschat.util.RestResponse;
 import lexek.wschat.util.RestResponseHandler;
-import org.apache.http.HttpEntity;
 import org.apache.http.client.HttpClient;
-import org.apache.http.client.entity.UrlEncodedFormEntity;
 import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.message.BasicNameValuePair;
 
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 
-public class GoodGameSocialAuthProvider implements SocialAuthProvider {
+public class VkAuthProvider implements SocialAuthProvider {
     private final String name;
     private final String clientId;
     private final String clientSecret;
@@ -28,7 +25,7 @@ public class GoodGameSocialAuthProvider implements SocialAuthProvider {
     private final HttpClient httpClient;
     private final SecureTokenGenerator secureTokenGenerator;
 
-    public GoodGameSocialAuthProvider(
+    public VkAuthProvider(
         String clientId,
         String clientSecret,
         String url,
@@ -47,10 +44,10 @@ public class GoodGameSocialAuthProvider implements SocialAuthProvider {
     @Override
     public SocialRedirect getRedirect() {
         String state = secureTokenGenerator.generateRandomToken(32).replace("=", "");
-        String result = "http://api2.goodgame.ru/oauth/authorize?response_type=code" +
+        String result = "https://oauth.vk.com/authorize?response_type=code" +
             "&client_id=" + UrlEscapers.urlPathSegmentEscaper().escape(clientId) +
             "&redirect_uri=" + UrlEscapers.urlPathSegmentEscaper().escape(url) +
-            "&scope=chat.token" +
+            "&display=page" +
             "&state=" + UrlEscapers.urlPathSegmentEscaper().escape(state) +
             "&prompt=consent" +
             "&access_type=offline";
@@ -59,70 +56,54 @@ public class GoodGameSocialAuthProvider implements SocialAuthProvider {
 
     @Override
     public SocialToken authenticate(String code) throws IOException {
-        HttpPost request = new HttpPost("http://api2.goodgame.ru/oauth");
-        HttpEntity entity = new UrlEncodedFormEntity(ImmutableList.of(
-            new BasicNameValuePair("client_secret", clientSecret),
-            new BasicNameValuePair("client_id", clientId),
-            new BasicNameValuePair("grant_type", "authorization_code"),
-            new BasicNameValuePair("redirect_uri", url),
-            new BasicNameValuePair("code", code)
-        ), CharsetUtil.UTF_8);
-        request.setEntity(entity);
+        String tokenUrl = "https://oauth.vk.com/access_token" +
+            "?client_id=" + UrlEscapers.urlPathSegmentEscaper().escape(clientId) +
+            "&client_secret=" + UrlEscapers.urlPathSegmentEscaper().escape(clientSecret) +
+            "&redirect_uri=" + UrlEscapers.urlPathSegmentEscaper().escape(url) +
+            "&code=" + UrlEscapers.urlPathSegmentEscaper().escape(code);
+        HttpGet request = new HttpGet(tokenUrl);
         request.setHeader(HttpHeaders.ACCEPT, "application/json");
         RestResponse response = httpClient.execute(request, RestResponseHandler.INSTANCE);
         JsonNode rootNode = response.getRootNode();
         if (response.isSuccess()) {
             String token = rootNode.get("access_token").asText();
-            String refreshToken = rootNode.get("refresh_token").asText();
             long expiresIn = TimeUnit.SECONDS.toMillis(rootNode.get("expires_in").asLong() - 60);
             return new SocialToken(
                 name,
                 token,
                 System.currentTimeMillis() + expiresIn,
-                refreshToken
+                null
             );
         } else {
-            throw new InvalidStateException(rootNode.get("detail").asText());
+            throw new InvalidStateException(rootNode.get("error_description").asText());
         }
     }
 
     @Override
     public SocialProfile getProfile(SocialToken token) throws IOException {
-        HttpGet request = new HttpGet("http://api2.goodgame.ru/info");
+        String profileUrl = "https://api.vk.com/method/users.get" +
+            "?fields=name,last_name,first_name" +
+            "&access_token=" + token.getToken();
+        HttpGet request = new HttpGet(profileUrl);
         request.setHeader(HttpHeaders.AUTHORIZATION, "Bearer " + token.getToken());
         request.setHeader(HttpHeaders.ACCEPT, "application/json");
-        JsonNode userData = httpClient.execute(request, JsonResponseHandler.INSTANCE);
-        JsonNode userNode = userData.get("user");
-        String externalName = userNode.get("username").asText().toLowerCase();
-        String externalId = userNode.get("user_id").asText();
+        JsonNode users = httpClient.execute(request, JsonResponseHandler.INSTANCE).get("response");
+        JsonNode userNode = users.get(0);
+        String firstName = userNode.get("first_name").asText();
+        String lastName = userNode.get("last_name").asText();
+        String externalName = firstName + " " + lastName;
+        String externalId = userNode.get("uid").asText();
         return new SocialProfile(externalId, name, externalName, null, token);
     }
 
     @Override
     public SocialToken refresh(SocialToken oldToken) throws IOException {
-        HttpPost request = new HttpPost("http://api2.goodgame.ru/oauth");
-        request.setHeader(HttpHeaders.ACCEPT, "application/json");
-        request.setEntity(new UrlEncodedFormEntity(ImmutableList.of(
-            new BasicNameValuePair("client_id", clientId),
-            new BasicNameValuePair("client_secret", clientSecret),
-            new BasicNameValuePair("refresh_token", oldToken.getRefreshToken()),
-            new BasicNameValuePair("grant_type", "refresh_token")
-        )));
-        JsonNode response = httpClient.execute(request, JsonResponseHandler.INSTANCE);
-        String token = response.get("access_token").asText();
-        String refreshToken = response.get("refresh_token").asText();
-        long expiresIn = TimeUnit.SECONDS.toMillis(response.get("expires_in").asLong() - 60);
-        return new SocialToken(
-            name,
-            token,
-            System.currentTimeMillis() + expiresIn,
-            refreshToken
-        );
+        throw new UnsupportedOperationException();
     }
 
     @Override
     public boolean needsRefreshing() {
-        return true;
+        return false;
     }
 
     @Override
